mysite/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
idna==2.10
polib==1.1.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
gunicorn==20.0.4
psycopg2-binary==2.9.6
django-two-factor-auth
phonenumbers
pyotp
pyjwt
web3
